Updated Discord invite. Made Readme more pleasant.
The readme had some minor typos. As well as some writing issues like sentences that trailed and an over-use of passive speech. Edited it to make it more appealing to a new-comer to read. Along with making it easier to edit by inserting line-breaks in markdown that don't affect the rendered page.