Added function/button to launch CP77 via Steam
Adds a function and a button to the shell next to the existing "Launch Game" button that runs the default steam id for Cyberpunk 2077. Not fully added to the ribbon view however as i don't know how to do that. Seemingly "RibbonView.g.i.cs" would have to be generated again. I hardly know what I'm doing so please be careful with this