fix!: Morphtarget Export Writes Correct Diff/Mapping Data
Vertex diff mappings are stored contiguously in the RE4 buffer, 2 bytes per.
Because the mapping count is for some reason stored as half the actual count,
it needs to be padded by an extra empty 2byte when the diff count is odd.

Tried to make the logic a bit more explicitly laid out, too.