fix!: Morphtarget Import Produces Correct, Functioning Morphs
Much of this is restructuring the logic and making it more explicit,
but the bookkeeping and data buffer construction now produce 1:1
results compared against vanilla. Should support:

1. Vanilla reimports with unchanged morphtargets
2. Modified vanilla
3. Custom morphtargets (although the *game* doesn't do anything with these)

BREAKING CHANGE: Relies on the new, fixed export logic. Will fail at
least on some previously exported (and thus broken) morphtargets. Might
be able to add fixes for some cases, but needs to be flagged to users.